requirements.txt
# Core Dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
click==8.1.7
rich==13.7.0

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1
asyncpg==0.29.0

# Redis & Queue
redis==5.0.1
celery==5.3.4

# Data Processing
pandas==2.1.4
numpy==1.26.2
pydantic==2.5.2
pydantic-settings==2.1.0

# Network Tools
scapy==2.5.0
netifaces==0.11.0
pyshark==0.6

# Monitoring & Logging
prometheus-client==0.19.0
structlog==23.2.0
python-json-logger==2.0.7

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
pytest-mock==3.12.0

# Development
python-dotenv==1.0.0
pyyaml==6.0.1
watchdog==3.0.0
black==23.12.0
ruff==0.1.8

# UI/Dashboard (optional)
streamlit==1.29.0
plotly==5.18.0

# Security
cryptography==41.0.7
paramiko==3.4.0  # For SSH connections